dev/backend/requirements.txt
accelerate==0.27.2
aiofiles==22.1.0
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.19.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bitsandbytes==0.42.0
bleach==6.0.0
blinker==1.7.0
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
-e git+https://github.com/facebookresearch/codellama@e66609cfbd73503ef25e597fd82c59084836155d#egg=codellama
comm==0.1.3
ConfigArgParse==1.5.3
datasets==2.17.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.6
docstring-parser==0.15
einops==0.7.0
et-xmlfile==1.1.0
executing==1.2.0
fairscale==0.4.13
fastjsonschema==2.16.3
filelock==3.12.0
fire==0.5.0
Flask==3.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.10.0
gevent==22.10.2
geventhttpclient==2.0.8
gguf==0.6.0
gql==3.5.0
graphql-core==3.2.3
greenlet==2.0.0.post0
huggingface-hub==0.20.3
idna==3.4
ipykernel==6.22.0
ipython==8.13.2
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
locust==2.13.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
msgpack==1.0.4
multidict==6.0.4
multiprocess==0.70.16
mysql-connector-python==8.3.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.4
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
peft==0.8.2
pexpect==4.8.0
pg==0.1
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.5.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.25.3
psutil==5.9.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycparser==2.21
Pygments==2.15.1
PyOpenGL==3.1.7
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==24.0.1
regex==2023.12.25
requests==2.28.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
roundrobin==0.0.4
safetensors==0.4.2
scipy==1.12.0
Send2Trash==1.8.2
sentencepiece==0.1.99
shtab==1.7.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.28
stack-data==0.6.2
sympy==1.12
termcolor==2.3.0
terminado==0.17.1
tinycss2==1.2.1
tokenizers==0.15.2
torch==2.1.1
tornado==6.3.1
tqdm==4.66.2
traitlets==5.9.0
transformers==4.38.1
trl==0.7.11
typing_extensions==4.10.0
tyro==0.7.3
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.12
virtualenv==20.23.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==3.0.1
xxhash==3.4.1
y-py==0.5.9
yarl==1.9.4
ypy-websocket==0.8.2
zope.event==4.5.0
zope.interface==5.5.1
